backend/requirements.txt
aiofiles==23.2.1
alembic==1.13.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
git-filter-repo==2.47.0
greenlet==3.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.5
MarkupSafe==2.1.3
orjson==3.9.10
packaging==23.2
pillow==10.4.0
pluggy==1.3.0
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.9.0
pypng==0.20220715.0
pytest==7.4.3
pytest-mock==3.12.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
qrcode==7.4.2
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.5.3
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2025.2
ujson==5.8.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
